+++ CHANGELOG.md
@@ -7,6 +7,11 @@
 - Add option for max number of files
 - Add store_container option to filepicker_field
 - Add open_to option to filepicker_field
+- Add filepicker_field_tag helper method
+
+### improvements
+
+- move from end-of-life multi_json to json
 
 ## 1.3.0 (December 31, 2014)